Fixed Triangle solution for overflowing arithmetics case (max int)

+++ src/main/java/world/coding/challenges/codility/Triangle.java
@@ -21,7 +21,7 @@ public class Triangle {
                 int a = A[p];
                 int b = A[p+1];
                 int c = A[p+2];
-                if (a+b>c){
+                if ((long)a+b>c){
                     output++;
                     break;
                 }
